requirements.txt
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.1
decorator==4.0.11
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.6.0
instagrapi==2.1.3
markdown-it-py==3.0.0
mdurl==0.1.2
moviepy==1.0.3
numpy==2.2.1
Pillow==9.5.0
proglog==0.1.10
pycryptodomex==3.21.0
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.19.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==0.7.0
rich==13.9.4
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
